Cleaning
* some unused structures and functions
* one errors on sys_recvfrom to run send_clientserver

+++ src/sys_recvfrom.c
@@ -50,7 +50,7 @@ void syscall_recvfrom_pre(pid_t pid, reg_s * reg, syscall_arg_u * sysarg, proces
 
   recvfrom_arg_t arg = &(sysarg->recvfrom);
 
-  if ( (int) reg->ret > 0) {
+  if ( reg->ret > 0) {
     fd_descriptor_t *file_desc = process_descriptor_get_fd(proc, arg->sockfd);
     file_desc->refcount++;
